// packages/gte-ts/src/internal/generated/types.gen.ts
export type OrderStatus = 'new' | 'partially_filled' | 'filled' | 'cancelled' | 'rejected' | 'expired' | 'replaced' | 'pending_new';

export type OrderType = 'market' | 'limit' | 'stop_limit' | 'stop_market' | 'liquidation';

export type PositionSide = 'long' | 'short';

export type Side = 'buy' | 'sell';

export type TimeInForce = 'gtc' | 'ioc' | 'fok' | 'gtd';

export type TokenType = 'crypto' | 'stablecoin' | 'equities' | 'commodities';

export type Tpsl = 'tp' | 'sl';

export type TradeDirection = 'open_long' | 'open_short' | 'long_to_short' | 'short_to_long' | 'close_long' | 'close_short';

export type TwapStatus = 'activated' | 'finished' | 'terminated' | 'error';

export type SetLeverageBody = {
    symbol: string;
    leverage: number;
    subaccountId: number;
    nonce: number;
    signature: string;
};

export type PerpPosition = {
    accountId: string;
    marketSymbol: string;
    size: string;
    positionValue: string;
    side: PositionSide;
    entryPrice: string;
    markPrice: string;
    liquidationPrice: string;
    leverage: number;
    margin: string;
    funding: string;
    isCross: boolean;
    subaccountId: number;
    unrealizedPnl: string;
    timestamp: string;
};
// ...
export type Candle = {
    /**
     * Candle open-time as epoch **milliseconds**.
     */
    timestamp: string;
    open: number;
    high: number;
    low: number;
    close: number;
    volume: number;
    tradesChecksum?: string | null;
};
// ...
export type FundingPayment = {
    marketSymbol: string;
    timestamp: string;
    side: PositionSide;
    payment: string;
    size: string;
    fundingRate: string;
    blockNumber?: string;
    txnIndex?: number;
};
// ...
export type GetBalancesResponse = {
    spot: Array<TokenBalance>;
    perps: Array<TokenBalance>;
    version: number;
};
// ...
export type HttpBook = {
    asks: Array<HttpBookLevel>;
    bids: Array<HttpBookLevel>;
    timestamp: string;
};

export type HttpBookLevel = {
    price: number;
    qty: string;
    numOrders: number;
};
// ...
export type MarketDataPerps = {
    openInterest: number;
    markPrice: number;
    indexPrice: number;
    midPrice: number;
    fundingRate: number;
    bestBid?: number | null;
    bestAsk?: number | null;
    bestBidVolume?: number | null;
    bestAskVolume?: number | null;
    impactPrice?: number | null;
    volume24h?: number | null;
    prevDayPrice?: number | null;
};
// ...
export type PerpOrder = {
    marketSymbol: string;
    timestamp: string;
    orderId: string;
    originalSize: string;
    side: Side;
    currentSize: string;
    orderValue: string;
    price: string;
    isReduceOnly: boolean;
    orderType: OrderType;
    status: OrderStatus;
    clientId?: string | null;
    triggerPrice?: string;
    tpsl?: Tpsl;
    leverage: string;
};
// ...
export type Token = {
    symbol: string;
    name: string;
    address?: string | null;
    decimals: number;
    logoUrl: string;
    tokenType: TokenType;
};

export type TokenBalance = {
    token?: null | Token;
    totalBalance: number;
    balanceUsd: number;
    freeCollateral: number;
    tradingAllowance: number;
};